# swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "nocodb",
        "version": "2.0"
    },
    "servers": [
        {
            "url": "http://localhost:8080"
        },
        {
            "url": "{customUrl}",
            "variables": {
                "customUrl": {
                    "default": "http://localhost:8080",
                    "description": "Provide custom nocodb app base url"
                }
            }
        }
    ],
    "paths": {
...
        "/api/v2/tables/m31ziu6jn3a3qs7/records": {
            "get": {
                "summary": "assets list",
                "operationId": "assets-db-table-row-list",
                "description": "List of all rows from assets table and response data fields can be filtered based on query params.",
                "tags": [
                    "assets"
                ],
                "parameters": [
                    {
                        "schema": {
                            "type": "string",
                            "enum": [
                                "vw1ukyvwx6v2ib2v"
                            ]
                        },
                        "description": "Allows you to fetch records that are currently visible within a specific view.\n\nViews:\n* vw1ukyvwx6v2ib2v - Default view",
                        "name": "viewId",
                        "in": "query",
                        "required": false
                    },
                    {
                        "schema": {
                            "type": "string"
                        },
                        "in": "query",
                        "name": "fields",
                        "description": "Array of field names or comma separated filed names to include in the response objects. In array syntax pass it like `fields[]=field1&fields[]=field2` or alternately `fields=field1,field2`."
                    },
                    {
                        "schema": {
                            "type": "string"
                        },
                        "in": "query",
                        "name": "sort",
                        "description": "Comma separated field names to sort rows, rows will sort in ascending order based on provided columns. To sort in descending order provide `-` prefix along with column name, like `-field`. Example : `sort=field1,-field2`"
                    },
                    {
                        "schema": {
                            "type": "string"
                        },
                        "in": "query",
                        "name": "where",
                        "description": "This can be used for filtering rows, which accepts complicated where conditions. For more info visit [here](https://docs.nocodb.com/developer-resources/rest-apis#comparison-operators). Example : `where=(field1,eq,value)`"
                    },
                    {
                        "schema": {
                            "type": "number",
                            "minimum": 1
                        },
                        "in": "query",
                        "name": "limit",
                        "description": "The `limit` parameter used for pagination, the response collection size depends on limit value with default value `25` and maximum value `1000`, which can be overridden by environment variables `DB_QUERY_LIMIT_DEFAULT` and `DB_QUERY_LIMIT_MAX` respectively.",
                        "example": 25
                    },
                    {
                        "schema": {
                            "type": "number",
                            "minimum": 0,
                            "maximum": 1
                        },
                        "in": "query",
                        "name": "shuffle",
                        "description": "The `shuffle` parameter used for pagination, the response will be shuffled if it is set to 1.",
                        "example": 0
                    },
                    {
                        "schema": {
                            "type": "number",
                            "minimum": 0
                        },
                        "in": "query",
                        "name": "offset",
                        "description": "The `offset` parameter used for pagination, the value helps to select collection from a certain index.",
                        "example": 0
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "list": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/assetsResponse"
                                            }
                                        },
                                        "PageInfo": {
                                            "$ref": "#/components/schemas/Paginated"
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "summary": "assets create",
                "description": "Insert a new row in table by providing a key value pair object where key refers to the column alias. All the required fields should be included with payload excluding `autoincrement` and column with default value.",
                "operationId": "assets-create",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/assetsResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "$ref": "#/components/responses/BadRequest"
                    }
                },
                "tags": [
                    "assets"
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "oneOf": [
                                    {
                                        "$ref": "#/components/schemas/assetsRequest"
                                    },
                                    {
                                        "type": "array",
                                        "items": {
                                            "$ref": "#/components/schemas/assetsRequest"
                                        }
                                    }
                                ]
                            }
                        }
                    }
                }
            },
            "patch": {
                "summary": "assets update",
                "operationId": "assets-update",
                "description": "Partial update row in table by providing a key value pair object where key refers to the column alias. You need to only include columns which you want to update.",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "$ref": "#/components/responses/BadRequest"
                    }
                },
                "tags": [
                    "assets"
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "oneOf": [
                                    {
                                        "$ref": "#/components/schemas/assetsRequest"
                                    },
                                    {
                                        "type": "array",
                                        "items": {
                                            "$ref": "#/components/schemas/assetsRequest"
                                        }
                                    }
                                ]
                            }
                        }
                    }
                }
            },
            "delete": {
                "summary": "assets delete",
                "operationId": "assets-delete",
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "tags": [
                    "assets"
                ],
                "description": "Delete a row by using the **primary key** column value.",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "oneOf": [
                                    {
                                        "$ref": "#/components/schemas/assetsIdRequest"
                                    },
                                    {
                                        "type": "array",
                                        "items": {
                                            "$ref": "#/components/schemas/assetsIdRequest"
                                        }
                                    }
                                ]
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "Paginated": {
                "title": "Paginated",
                "type": "object",
                "properties": {
                    "pageSize": {
                        "type": "integer"
                    },
                    "totalRows": {
                        "type": "integer"
                    },
                    "isFirstPage": {
                        "type": "boolean"
                    },
                    "isLastPage": {
                        "type": "boolean"
                    },
                    "page": {
                        "type": "number"
                    }
                }
            },
...
            "assetsResponse": {
                "title": "assets Response",
                "type": "object",
                "description": "",
                "x-internal": false,
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "name": {
                        "type": "string"
                    },
                    "type": {
                        "type": "string"
                    },
                    "s3_key": {
                        "type": "string"
                    },
                    "content_type": {
                        "type": "string"
                    },
                    "file_size": {
                        "type": "integer"
                    },
                    "metadata": {
                        "type": "object"
                    },
                    "created_at": {
                        "type": "string"
                    }
                }
            },
            "assetsRequest": {
                "title": "assets Request",
                "type": "object",
                "description": "",
                "x-internal": false,
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "name": {
                        "type": "string"
                    },
                    "type": {
                        "type": "string"
                    },
                    "s3_key": {
                        "type": "string"
                    },
                    "content_type": {
                        "type": "string"
                    },
                    "file_size": {
                        "type": "integer"
                    },
                    "metadata": {
                        "type": "object"
                    }
                }
            },
            "assetsIdRequest": {
                "title": "assets Id Request",
                "type": "object",
                "description": "",
                "x-internal": false,
                "properties": {
                    "id": {
                        "type": "integer"
                    }
                }
            },
...
        },
        "securitySchemes": {
            "xcToken": {
                "type": "apiKey",
                "in": "header",
                "name": "xc-token",
                "description": "API token"
            }
        },
        "responses": {
            "BadRequest": {
                "description": "BadRequest",
                "content": {
                    "application/json": {
                        "schema": {
                            "type": "object",
                            "properties": {
                                "msg": {
                                    "type": "string",
                                    "x-stoplight": {
                                        "id": "p9mk4oi0hbihm"
                                    },
                                    "example": "BadRequest [Error]: <ERROR MESSAGE>"
                                }
                            },
                            "required": [
                                "msg"
                            ]
                        },
                        "examples": {
                            "Example 1": {
                                "value": {
                                    "msg": "BadRequest [Error]: <ERROR MESSAGE>"
                                }
                            }
                        }
                    }
                },
                "headers": {}
            }
        }
    },
    "security": [
        {
            "xcToken": []
        }
    ]
}